+++ iam_python_sdk/__init__.py
@@ -13,12 +13,14 @@
 # limitations under the License.
 
 from iam_python_sdk.client import NewDefaultClient
+from iam_python_sdk.async_client import NewAsyncClient
 from iam_python_sdk.config import Config
 
 __author__ = """Analytics AccelByte"""
 __email__ = 'justice-analytics-team@accelbyte.net'
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 __all__ = [
     'NewDefaultClient',
+    'NewAsyncClient',
     'Config'
 ]